made changes on patient request page.
