+++ interface/midocs/createpatient.php
@@ -1,5 +1,4 @@
 <?php
-session_start();
 
 /**
  * Login screen.
