+++ interface/midocs/createprovider.php
@@ -1,6 +1,4 @@
 <?php
-session_start();
-
 /**
  * Login screen.
  *
