+++ patches/0007-make-design-changes-in-input-fields.patch
@@ -1,0 +1,41 @@
+From 454068f1705a536d32e3309af68560cb5500312d Mon Sep 17 00:00:00 2001
+From: Aiyash Ahmed <aiyashahmed96@gmail.com>
+Date: Wed, 18 Nov 2020 14:37:38 +0530
+Subject: [PATCH 7/7] make design changes in input fields
+
+---
+ interface/midocs/createprovider.php | 6 +++---
+ 1 file changed, 3 insertions(+), 3 deletions(-)
+
+diff --git a/interface/midocs/createprovider.php b/interface/midocs/createprovider.php
+index 3df6abf45..bd0cb6ffe 100644
+--- a/interface/midocs/createprovider.php
++++ b/interface/midocs/createprovider.php
+@@ -171,21 +171,21 @@ $url = substr($url, 0, strpos($url, "interface"));
+             </div>
+             <div class="form-group row">
+                 <label for="npi" class="col-sm-2">Individual NPI</label>
+-                <div class="col-sm-4">
++                <div class="col-sm-6">
+                     <input type="text" class="form-control" name="npi">
+                     <label class="badge text-danger mt-1 error-field" id="npi"></label>
+                 </div>
+             </div>
+             <div class="form-group row">
+                 <label for="email" class="col-sm-2">E-mail</label>
+-                <div class="col-sm-4">
++                <div class="col-sm-8">
+                     <input type="text" class="form-control" name="email">
+                     <label class="badge text-danger mt-1 error-field" id="email"></label>
+                 </div>
+             </div>
+             <div class="form-group row">
+                 <label for="con_email" class="col-sm-2">Confirm E-mail</label>
+-                <div class="col-sm-4">
++                <div class="col-sm-8">
+                     <input type="email" class="form-control" name="con_email">
+                     <label class="badge text-danger mt-1 error-field" id="con_email"></label>
+                 </div>
+-- 
+2.27.0.windows.1
+
